Fixes compilation on some platforms
At least on avr32 Linux with gcc, there was missing casts to a fxmin parameters.

+++ faxd/faxQueueApp.c++
@@ -1077,7 +1077,7 @@ faxQueueApp::preparePageChop(const FaxRequest& req,
 	if (dec.getLastBlanks() > minRows)
 	{
 	    pagehandling.append(fxStr::format("Z%04x",
-		fxmin((unsigned)0xFFFF, stripSize - (dec.getEndOfPage() - data))));
+		fxmin((u_int)0xFFFF, (u_int)(stripSize - (dec.getEndOfPage() - data)))));
 	}
     }
     delete [] data;
